src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
